fix DB-related crashes as well + some refatoring
The logic to resolve the DB path also caused crashes if the program dir
is read-only -> fixed.
Also moved some generic functions to CoreUtils to avoid duplication and
improve separation of concerns.
