+++ VDF.Core/Utils/Logger.cs
@@ -13,7 +13,6 @@
 //     along with VideoDuplicateFinder.  If not, see <http://www.gnu.org/licenses/>.
 // */
 //
-using System.Runtime.InteropServices;
 
 namespace VDF.Core.Utils {
 	public sealed class Logger {
@@ -39,26 +38,7 @@ namespace VDF.Core.Utils {
 			if (CoreUtils.CanWriteToDirectory(CoreUtils.CurrentFolder))
 				return CoreUtils.CurrentFolder;
 
-			return GetDefaultStateFolder();
-		}
-		static string GetDefaultStateFolder() {
-			string? baseFolder;
-			if (CoreUtils.IsWindows) {
-				// LocalApplicationData = %LOCALAPPDATA% (local, non-roaming — appropriate for state/logs)
-				baseFolder = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-			}
-			else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX)) {
-				baseFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "Library", "Application Support");
-			}
-			else {
-				baseFolder = Environment.GetEnvironmentVariable("XDG_STATE_HOME");
-				if (string.IsNullOrWhiteSpace(baseFolder))
-					baseFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".local", "state");
-			}
-
-			var stateFolder = Path.Combine(baseFolder, "VDF");
-			Directory.CreateDirectory(stateFolder);
-			return stateFolder;
+			return CoreUtils.GetDefaultStateFolder();
 		}
 	}
